requirements.txt
torch==2.1.0+cu118
torchvision==0.16.0+cu118
--extra-index-url https://download.pytorch.org/whl/cu118
matplotlib
tqdm
pillow
tkinterdnd2